Add form, move utils, and examples loader
